+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_04_22_193758) do
+ActiveRecord::Schema.define(version: 2021_04_22_225248) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -119,6 +119,7 @@ ActiveRecord::Schema.define(version: 2021_04_22_193758) do
     t.boolean "employed_as_supply_teacher"
     t.boolean "has_entire_term_contract"
     t.boolean "employed_directly"
+    t.boolean "subject_to_disciplinary_action"
   end
 
   create_table "local_authorities", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|