+++ spec/models/early_career_payments/eligibility_answers_presenter_spec.rb
@@ -5,7 +5,8 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
     {
       nqt_in_academic_year_after_itt: true,
       employed_as_supply_teacher: false,
-      employed_directly: true
+      employed_directly: true,
+      subject_to_disciplinary_action: false
     }
   end
   let(:eligibility) { claim.eligibility }
@@ -16,7 +17,8 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
   it "returns an array of questions and answers to be presented to the user for checking" do
     expected_answers = [
       [I18n.t("early_career_payments.questions.nqt_in_academic_year_after_itt"), "Yes", "nqt-in-academic-year-after-itt"],
-      [I18n.t("early_career_payments.questions.employed_as_supply_teacher"), "No", "supply-teacher"]
+      [I18n.t("early_career_payments.questions.employed_as_supply_teacher"), "No", "supply-teacher"],
+      [I18n.t("early_career_payments.questions.disciplinary_action"), "No", "disciplinary-action"]
     ]
 
     expect(presenter.answers).to eq(expected_answers)
@@ -28,7 +30,8 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
         nqt_in_academic_year_after_itt: true,
         employed_as_supply_teacher: true,
         has_entire_term_contract: true,
-        employed_directly: true
+        employed_directly: true,
+        subject_to_disciplinary_action: false
       }
     end
 
@@ -37,7 +40,8 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
         [I18n.t("early_career_payments.questions.nqt_in_academic_year_after_itt"), "Yes", "nqt-in-academic-year-after-itt"],
         [I18n.t("early_career_payments.questions.employed_as_supply_teacher"), "Yes", "supply-teacher"],
         [I18n.t("early_career_payments.questions.has_entire_term_contract"), "Yes", "entire-term-contract"],
-        [I18n.t("early_career_payments.questions.employed_directly"), "Yes", "employed-directly"]
+        [I18n.t("early_career_payments.questions.employed_directly"), "Yes", "employed-directly"],
+        [I18n.t("early_career_payments.questions.disciplinary_action"), "No", "disciplinary-action"]
       ]
 
       expect(presenter.answers).to eq(expected_answers)